Rename dump.rs to dump_util.rs; remove some TODOs
