+++ crates/starknet-devnet-server/src/api/http/endpoints/dump_load.rs
@@ -6,7 +6,7 @@ use crate::api::http::error::HttpApiError;
 use crate::api::http::models::{DumpPath, DumpResponseBody};
 use crate::api::http::{HttpApiHandler, HttpApiResult};
 use crate::api::Api;
-use crate::dump::dump_events;
+use crate::dump_util::dump_events;
 
 pub async fn dump(
     State(state): State<HttpApiHandler>,
