+++ crates/starknet-devnet-server/src/api/mod.rs
@@ -7,7 +7,7 @@ use std::sync::Arc;
 use starknet_core::starknet::Starknet;
 use tokio::sync::Mutex;
 
-use crate::dump::DumpEvent;
+use crate::dump_util::DumpEvent;
 
 /// Data that can be shared between threads with read write lock access
 /// Whatever needs to be accessed as information outside of Starknet could be added to this struct
