+++ crates/starknet-devnet-server/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod api;
 mod config;
-pub mod dump;
+pub mod dump_util;
 pub mod error;
 pub mod rpc_core;
 /// handlers for axum server
